# Reimplement adding pieces and counting merges to speed up generate_early