--- a/worlds/jigsaw/Rules.py
+++ b/worlds/jigsaw/Rules.py
@@ -14,9 +14,6 @@ class PuzzleBoard:
     Optimized for efficient adding pieces to the board and calculating the number of new merges that would be made when
     adding a piece.
 
-    The pieces in each group are not tracked because this makes adding pieces more expensive, this makes piece removal
-    more expensive, however, the removal of pieces is not expected to occur often.
-
     A board stores a group ID value for each piece placed on the board, where `None` indicates an empty space on the
     board.
     _0____3_
@@ -40,22 +37,13 @@ class PuzzleBoard:
     __1_22_5
 
     C)
-    When a piece is placed that would connect to multiple existing groups, one way to handle this is to set all the
-    pieces in those groups to the group ID of the largest group.
+    When a piece is placed that would connect to multiple existing groups, the largest of the groups is found and all
+    pieces in the smaller groups are added to the largest group.
+    The group IDs of the smaller groups can then be re-used
     _0____3_     _0____3_
     ____4_33     ____4_33
     __1____X <-- __1____3 <--
     __1_22_5     __1_22_3 <--
-    However, this gets expensive to do as the groups get larger. Instead, a dictionary is used that maps each group ID
-    already on the board to a 'real' group ID:
-    _0____3_     _0____3_
-    ____4_33     ____4_33
-    __1____X <-- __1____3 <--
-    __1_22_5     __1_22_5
-    group_to_real_group[5] = 3
-    real_group_to_groups[3].append(5)
-    This means that other operations that need to check what group a piece is in need to get the group ID as stored on
-    the board, but then look up the real group ID with `real_group_id = group_to_real_group[group_id_from_board]`.
     """
 
     # The puzzle board itself. A 1D list used to represent a 2D board.
@@ -64,10 +52,8 @@ class PuzzleBoard:
     adjacent_pieces: tuple[tuple[int, ...], ...]
     # The count of merged groups of connected pieces.
     merges_count: int
-    # group ID -> real group ID, used like a dict[int, int]
-    group_to_real_group: list[int]
-    # real group ID -> group IDs mapped to this real group ID
-    real_group_to_groups: dict[int, list[int]]
+    # group ID -> piece indices in the group.
+    groups: dict[int, list[int]]
 
     # Unused group IDs that newly added pieces can be assigned to if they do not merge into an existing group.
     _unused_ids: list[int]
@@ -96,10 +82,8 @@ class PuzzleBoard:
                 piece_connections.append(i + width)
             adjacent_pieces.append(tuple(piece_connections))
         self.adjacent_pieces = tuple(adjacent_pieces)
-
         self.merges_count = 0
-        self.group_to_real_group = [-1] * len(self._unused_ids)
-        self.real_group_to_groups = {}
+        self.groups = {}
 
     def add_piece(self, piece_index: int):
         """
@@ -110,61 +94,64 @@ class PuzzleBoard:
         board = self.board
 
         # Get all adjacent group IDs.
-        found_groups = {board[connection] for connection in self.adjacent_pieces[piece_index]}
+        found_groups: set[int] = {board[connection] for connection in self.adjacent_pieces[piece_index]}  # type: ignore
         # Empty spaces on the board are set to `None`.
-        found_groups.discard(None)
+        found_groups.discard(None)  # type: ignore
 
         num_adjacent_groups = len(found_groups)
         if num_adjacent_groups == 0:
             # Isolated piece, give it a new ID
             new_id = self._unused_ids.pop()
             board[piece_index] = new_id
-            self.group_to_real_group[new_id] = new_id
-            self.real_group_to_groups[new_id] = [new_id]
+            self.groups[new_id] = [piece_index]
         elif num_adjacent_groups == 1:
             # Only one connecting group
             self.merges_count += 1
-            board[piece_index] = next(iter(found_groups))
+            found_group = next(iter(found_groups))
+            board[piece_index] = found_group
+            self.groups[found_group].append(piece_index)
         else:
-            # Multiple connecting groups, determine the real groups.
-            group_to_real_group = self.group_to_real_group
-            real_groups = {group_to_real_group[found_group] for found_group in found_groups}
-            num_real_groups = len(real_groups)
-            self.merges_count += num_real_groups
-            if num_real_groups == 1:
-                board[piece_index] = next(iter(real_groups))
-            else:
-                # Pick the first real group as the new real group.
-                real_groups_iter = iter(real_groups)
-                new_real_group = next(real_groups_iter)
-                board[piece_index] = new_real_group
+            # Multiple connecting groups
+            self.merges_count += num_adjacent_groups
+            groups = self.groups
 
-                # Re-map all groups mapped to the remaining real groups to the new real group.
-                real_group_to_groups = self.real_group_to_groups
-                groups_mapped_to_real_group = real_group_to_groups[new_real_group]
-                for other_real_group in real_groups_iter:
-                    groups = real_group_to_groups[other_real_group]
-                    groups_mapped_to_real_group.extend(groups)
-                    del real_group_to_groups[other_real_group]
-                    for g in groups:
-                        group_to_real_group[g] = new_real_group
+            groups_iter = iter(found_groups)
+
+            first_group_id: int = next(groups_iter)
+            pieces_in_group = groups[first_group_id]
+            groups_and_ids = [(first_group_id, pieces_in_group)]
+
+            # Find the largest of the groups
+            largest_group_id = first_group_id
+            largest_pieces_in_group = pieces_in_group
+            largest_group_size = len(pieces_in_group)
+            for group_id in groups_iter:
+                pieces_in_group = groups[group_id]
+                group_size = len(pieces_in_group)
+                groups_and_ids.append((group_id, pieces_in_group))
+                if group_size > largest_group_size:
+                    largest_group_size = group_size
+                    largest_group_id = group_id
+                    largest_pieces_in_group = pieces_in_group
+
+            board[piece_index] = largest_group_id
+            largest_pieces_in_group.append(piece_index)
+
+            # Add the pieces in the smaller groups into the largest group.
+            unused_ids = self._unused_ids
+            for group_id, pieces_in_group in groups_and_ids:
+                if pieces_in_group is largest_pieces_in_group:
+                    continue
+                largest_pieces_in_group.extend(pieces_in_group)
+                del groups[group_id]
+                unused_ids.append(group_id)
+                for piece_idx in pieces_in_group:
+                    board[piece_idx] = largest_group_id
 
     def get_merges_from_adding_piece(self, piece_idx: int):
         """Get the number of merges that would be made by adding a piece."""
-
         # Get all adjacent group IDs.
         board = self.board
         found_groups = {board[connection] for connection in self.adjacent_pieces[piece_idx]}
         # Empty spaces on the board are set to `None`.
-        found_groups.discard(None)
-
-        num_found = len(found_groups)
-        if num_found == 0:
-            return 0
-        elif num_found == 1:
-            # Only one connecting group
-            return 1
-        else:
-            # Multiple connecting groups, return the number of real groups.
-            group_to_real_group = self.group_to_real_group
-            return len({group_to_real_group[found_group] for found_group in found_groups})
+        return len(found_groups) - 1 if None in found_groups else len(found_groups)