--- a/worlds/jigsaw/Rules.py
+++ b/worlds/jigsaw/Rules.py
@@ -14,7 +14,7 @@ class PuzzleBoard:
     Optimized for efficient adding pieces to the board and calculating the number of new merges that would be made when
     adding a piece.
 
-    A board stores a group ID value for each piece placed on the board, where `None` indicates an empty space on the
+    A board stores a cluster ID value for each piece placed on the board, where `None` indicates an empty space on the
     board.
     _0____3_
     ____4_3_
@@ -23,23 +23,23 @@ class PuzzleBoard:
 
     Piece placement behaviours:
     A)
-    A newly placed piece that does not connect to any existing placed pieces is given a new unique group ID.
+    A newly placed piece that does not connect to any existing placed pieces is given a new unique cluster ID.
     _0____3_
     ____4_3_
     __1_____
     __1_22_5 <--
 
     B)
-    A newly placed piece that connects to only a single group ID gains the group ID of the connecting piece.
+    A newly placed piece that connects to only a single piece gains the cluster ID of the connecting piece.
     _0____3_
     ____4_33 <--
     __1_____
     __1_22_5
 
     C)
-    When a piece is placed that would connect to multiple existing groups, the largest of the groups is found and all
-    pieces in the smaller groups are added to the largest group.
-    The group IDs of the smaller groups can then be re-used
+    When a piece is placed that would connect to multiple existing clusters of pieces, the largest of the clusters is
+    found and all pieces in the smaller clusters are added to the largest cluster.
+    The cluster IDs of the smaller clusters can then be re-used later on.
     _0____3_     _0____3_
     ____4_33     ____4_33
     __1____X <-- __1____3 <--
@@ -50,12 +50,12 @@ class PuzzleBoard:
     board: list[int | None]
     # A lookup of the adjacent pieces of each piece, used like a dict[int, tuple[int, ...]].
     adjacent_pieces: tuple[tuple[int, ...], ...]
-    # The count of merged groups of connected pieces.
+    # The count of merged clusters of connected pieces.
     merges_count: int
-    # group ID -> piece indices in the group.
-    groups: dict[int, list[int]]
+    # cluster ID -> piece indices in the cluster.
+    clusters: dict[int, list[int]]
 
-    # Unused group IDs that newly added pieces can be assigned to if they do not merge into an existing group.
+    # Unused cluster IDs that newly added pieces can be assigned to if they do not merge into an existing cluster.
     _unused_ids: list[int]
 
     def __init__(self, width: int, height: int):
@@ -83,7 +83,7 @@ class PuzzleBoard:
             adjacent_pieces.append(tuple(piece_connections))
         self.adjacent_pieces = tuple(adjacent_pieces)
         self.merges_count = 0
-        self.groups = {}
+        self.clusters = {}
 
     def add_piece(self, piece_index: int):
         """
@@ -93,65 +93,66 @@ class PuzzleBoard:
         """
         board = self.board
 
-        # Get all adjacent group IDs.
-        found_groups: set[int] = {board[connection] for connection in self.adjacent_pieces[piece_index]}  # type: ignore
+        # Get all adjacent cluster IDs.
+        # Use incorrect typing to begin with because it is more efficient to remove `None` this way.
+        found_clusters: set[int] = {board[connection] for connection in self.adjacent_pieces[piece_index]}  # type: ignore
         # Empty spaces on the board are set to `None`.
-        found_groups.discard(None)  # type: ignore
+        found_clusters.discard(None)  # type: ignore
 
-        num_adjacent_groups = len(found_groups)
-        if num_adjacent_groups == 0:
+        num_adjacent_clusters = len(found_clusters)
+        if num_adjacent_clusters == 0:
             # Isolated piece, give it a new ID
             new_id = self._unused_ids.pop()
             board[piece_index] = new_id
-            self.groups[new_id] = [piece_index]
-        elif num_adjacent_groups == 1:
-            # Only one connecting group
+            self.clusters[new_id] = [piece_index]
+        elif num_adjacent_clusters == 1:
+            # Only one connecting cluster
             self.merges_count += 1
-            found_group = next(iter(found_groups))
-            board[piece_index] = found_group
-            self.groups[found_group].append(piece_index)
+            found_cluster = next(iter(found_clusters))
+            board[piece_index] = found_cluster
+            self.clusters[found_cluster].append(piece_index)
         else:
-            # Multiple connecting groups
-            self.merges_count += num_adjacent_groups
-            groups = self.groups
+            # Multiple connecting clusters
+            self.merges_count += num_adjacent_clusters
+            clusters = self.clusters
 
-            groups_iter = iter(found_groups)
+            clusters_iter = iter(found_clusters)
 
-            first_group_id: int = next(groups_iter)
-            pieces_in_group = groups[first_group_id]
-            groups_and_ids = [(first_group_id, pieces_in_group)]
+            first_cluster_id: int = next(clusters_iter)
+            pieces_in_cluster = clusters[first_cluster_id]
+            clusters_and_ids = [(first_cluster_id, pieces_in_cluster)]
 
-            # Find the largest of the groups
-            largest_group_id = first_group_id
-            largest_pieces_in_group = pieces_in_group
-            largest_group_size = len(pieces_in_group)
-            for group_id in groups_iter:
-                pieces_in_group = groups[group_id]
-                group_size = len(pieces_in_group)
-                groups_and_ids.append((group_id, pieces_in_group))
-                if group_size > largest_group_size:
-                    largest_group_size = group_size
-                    largest_group_id = group_id
-                    largest_pieces_in_group = pieces_in_group
+            # Find the largest of the clusters
+            largest_cluster_id = first_cluster_id
+            largest_pieces_in_cluster = pieces_in_cluster
+            largest_cluster_size = len(pieces_in_cluster)
+            for cluster_id in clusters_iter:
+                pieces_in_cluster = clusters[cluster_id]
+                cluster_size = len(pieces_in_cluster)
+                clusters_and_ids.append((cluster_id, pieces_in_cluster))
+                if cluster_size > largest_cluster_size:
+                    largest_cluster_size = cluster_size
+                    largest_cluster_id = cluster_id
+                    largest_pieces_in_cluster = pieces_in_cluster
 
-            board[piece_index] = largest_group_id
-            largest_pieces_in_group.append(piece_index)
+            board[piece_index] = largest_cluster_id
+            largest_pieces_in_cluster.append(piece_index)
 
-            # Add the pieces in the smaller groups into the largest group.
+            # Add the pieces in the smaller clusters into the largest cluster.
             unused_ids = self._unused_ids
-            for group_id, pieces_in_group in groups_and_ids:
-                if pieces_in_group is largest_pieces_in_group:
+            for cluster_id, pieces_in_cluster in clusters_and_ids:
+                if pieces_in_cluster is largest_pieces_in_cluster:
                     continue
-                largest_pieces_in_group.extend(pieces_in_group)
-                del groups[group_id]
-                unused_ids.append(group_id)
-                for piece_idx in pieces_in_group:
-                    board[piece_idx] = largest_group_id
+                largest_pieces_in_cluster.extend(pieces_in_cluster)
+                del clusters[cluster_id]
+                unused_ids.append(cluster_id)
+                for piece_idx in pieces_in_cluster:
+                    board[piece_idx] = largest_cluster_id
 
     def get_merges_from_adding_piece(self, piece_idx: int):
         """Get the number of merges that would be made by adding a piece."""
-        # Get all adjacent group IDs.
+        # Get all adjacent cluster IDs.
         board = self.board
-        found_groups = {board[connection] for connection in self.adjacent_pieces[piece_idx]}
+        found_clusters = {board[connection] for connection in self.adjacent_pieces[piece_idx]}
         # Empty spaces on the board are set to `None`.
-        return len(found_groups) - 1 if None in found_groups else len(found_groups)
+        return len(found_clusters) - 1 if None in found_clusters else len(found_clusters)