## What is this fixing or adding?

Speeds up generate_early by reimplementing the puzzle data structure into a structure that is fast to add pieces to and fast to check for how many merges would be made if a piece was added.

This change has a particularly large impact on least merges possible order and to a lesser extent every piece fits order. Random order is affected the least.

The new implementation is backed by a list, so is zero-indexed, whereas the original code was one-indexed. The original code that adds pieces to the puzzle, and counts how many merges would be made by adding a piece, now has to subtract 1 when calling the new code.

This supersedes #2 because this is much more peformant.

The ability to remove pieces from a board is unused, but has been implemented due to there being an existing implementation to remove pieces from the original puzzle data structure.

## How was this tested?
Generations with a fixed seed with a 1000 piece random order yaml, a 1000 piece least merges order yaml, and a 1000 piece every piece fits order yaml produced the same result before and after, but with better generate_early() performance. The function, that prints how long `generate_early` took, was modified to always print the duration.

`python -O .\Generate.py --skip_output --seed 1`:
Example before:
```
Took 0.5239 seconds in JigsawWorld.generate_early for player 1, named EveryPieceFits.
Took 14.8665 seconds in JigsawWorld.generate_early for player 2, named LeastMerges.
Took 0.0839 seconds in JigsawWorld.generate_early for player 3, named RandomOrder.

Creating MultiWorld.
Creating Items.
Calculating Access Rules.
Running Item Plando.
Running Pre Main Fill.
Filling the multiworld with 2994 items.
Current fill step (Progression) at 1000/2641 items placed.
Current fill step (Progression) at 2000/2641 items placed.
Current fill step (Progression) at 2641/2641 items placed.
Skipping multiworld progression balancing.
Done. Skipped output/spoiler generation. Total Time: 24.232174199998553
```

Example after:
```
Took 0.0567 seconds in JigsawWorld.generate_early for player 1, named EveryPieceFits.
Took 0.1119 seconds in JigsawWorld.generate_early for player 2, named LeastMerges.
Took 0.0246 seconds in JigsawWorld.generate_early for player 3, named RandomOrder.

Creating MultiWorld.
Creating Items.
Calculating Access Rules.
Running Item Plando.
Running Pre Main Fill.
Filling the multiworld with 2994 items.
Current fill step (Progression) at 1000/2641 items placed.
Current fill step (Progression) at 2000/2641 items placed.
Current fill step (Progression) at 2641/2641 items placed.
Skipping multiworld progression balancing.
Done. Skipped output/spoiler generation. Total Time: 8.896331800002372
```
[Players.zip](https://github.com/user-attachments/files/19612201/Players.zip)