@spinerak 2025-04-06
Getting it back from 15 to 0.11 seconds is absolutely nuts! 
I have also done a quick test to verify the puzzle outcomes are the same, and they were.
I've browsed through the code and the general idea is clear to me, but I haven't looked into it in detail. Good enough for me!

Much appreciated!